Crate coins_bip39

The bip39 crate is heavily inspired by and reuses code from Wagyu under the MIT license. The difference being, the underlying extended private keys are generated using the bip32 crate, that depends on k256 instead of libsecp256k1.

Re-exports

pub use self::mnemonic::*;

pub use self::wordlist::*;

Modules

mnemonic

Mnemonic phrases

wordlist

Wordlists