coins_bip39::wordlist
Trait Wordlist

pub trait Wordlist {
    // Required method
    fn get_all() -> &'static [&'static str];

    // Provided methods
    fn get(index: usize) -> Result<&'static str, WordlistError> { ... }
    fn get_index(word: &str) -> Result<usize, WordlistError> { ... }
}

The Wordlist trait that every language’s wordlist must implement.

Required Methods

fn get_all() -> &'static [&'static str]

Returns the word list as a string.

Implementor’s note: this MUST be sorted

Provided Methods

fn get(index: usize) -> Result<&'static str, WordlistError>

Returns the word of a given index from the word list.

fn get_index(word: &str) -> Result<usize, WordlistError>

Returns the index of a given word from the word list.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.